+++ examples/relationship_one_to_one_models_auto.py
@@ -54,12 +54,12 @@ class TPayInfo(typing.Protocol):
 PayInfo: TPayInfo = models.PayInfo  # type: ignore
 
 
-class EmployeeDict(typing.TypedDict, total=True):
-    """TypedDict for properties that are required."""
+class EmployeeDict(typing.TypedDict, total=False):
+    """TypedDict for properties that are not required."""
 
-    id: int
-    name: str
-    pay_info: "PayInfoDict"
+    id: typing.Optional[int]
+    name: typing.Optional[str]
+    pay_info: typing.Optional["PayInfoDict"]
 
 
 class TEmployee(typing.Protocol):
@@ -71,16 +71,26 @@ class TEmployee(typing.Protocol):
     query: orm.Query
 
     # Model properties
-    id: int
-    name: str
-    pay_info: "TPayInfo"
+    id: typing.Optional[int]
+    name: typing.Optional[str]
+    pay_info: typing.Optional["TPayInfo"]
 
-    def __init__(self, id: int, name: str, pay_info: "TPayInfo") -> None:
+    def __init__(
+        self,
+        id: typing.Optional[int] = None,
+        name: typing.Optional[str] = None,
+        pay_info: typing.Optional["TPayInfo"] = None,
+    ) -> None:
         """Construct."""
         ...
 
     @classmethod
-    def from_dict(cls, id: int, name: str, pay_info: "PayInfoDict") -> "TEmployee":
+    def from_dict(
+        cls,
+        id: typing.Optional[int] = None,
+        name: typing.Optional[str] = None,
+        pay_info: typing.Optional["PayInfoDict"] = None,
+    ) -> "TEmployee":
         """Construct from a dictionary (eg. a POST payload)."""
         ...