+++ examples/unique_constraint_models_auto.py
@@ -9,19 +9,13 @@ from sqlalchemy import orm
 from open_alchemy import models
 
 
-class _EmployeeDictBase(typing.TypedDict, total=True):
-    """TypedDict for properties that are required."""
-
-    id: int
-    name: str
-    division: str
-
-
-class EmployeeDict(_EmployeeDictBase, total=False):
+class EmployeeDict(typing.TypedDict, total=False):
     """TypedDict for properties that are not required."""
 
+    id: typing.Optional[int]
+    name: typing.Optional[str]
     address: typing.Optional[str]
-    salary: typing.Optional[float]
+    division: typing.Optional[str]
 
 
 class TEmployee(typing.Protocol):
@@ -33,19 +27,17 @@ class TEmployee(typing.Protocol):
     query: orm.Query
 
     # Model properties
-    id: int
-    name: str
+    id: typing.Optional[int]
+    name: typing.Optional[str]
     address: typing.Optional[str]
-    division: str
-    salary: typing.Optional[float]
+    division: typing.Optional[str]
 
     def __init__(
         self,
-        id: int,
-        name: str,
-        division: str,
+        id: typing.Optional[int] = None,
+        name: typing.Optional[str] = None,
         address: typing.Optional[str] = None,
-        salary: typing.Optional[float] = None,
+        division: typing.Optional[str] = None,
     ) -> None:
         """Construct."""
         ...
@@ -53,11 +45,10 @@ class TEmployee(typing.Protocol):
     @classmethod
     def from_dict(
         cls,
-        id: int,
-        name: str,
-        division: str,
+        id: typing.Optional[int] = None,
+        name: typing.Optional[str] = None,
         address: typing.Optional[str] = None,
-        salary: typing.Optional[float] = None,
+        division: typing.Optional[str] = None,
     ) -> "TEmployee":
         """Construct from a dictionary (eg. a POST payload)."""
         ...