+++ tests/open_alchemy/conftest.py
@@ -4,8 +4,6 @@
 from unittest import mock
 
 import pytest
-import sqlalchemy
-from sqlalchemy import orm
 from sqlalchemy.ext import declarative
 
 import open_alchemy
@@ -69,18 +67,6 @@ def _mocked_handle_column(mocked_handle_column):
     return mocked_handle_column
 
 
-@pytest.fixture(scope="function", params=["sqlite:///:memory:"])
-def engine(request):
-    """Creates a sqlite engine."""
-    return sqlalchemy.create_engine(request.param)
-
-
-@pytest.fixture(scope="function")
-def sessionmaker(engine):
-    """Creates a sqlite session."""
-    return orm.sessionmaker(bind=engine)
-
-
 @pytest.fixture(scope="function")
 def args():
     """Arguments."""
@@ -127,26 +113,6 @@ def mocked_declarative_base(monkeypatch):
     return mock_declarative_base
 
 
-@pytest.fixture(autouse=True)
-def cleanup_models():
-    """Remove any new attributes on open_alchemy.models."""
-    for key in set(models.__dict__.keys()):
-        if key.startswith("__"):
-            continue
-        if key.endswith("__"):
-            continue
-        delattr(models, key)
-
-    yield
-
-    for key in set(models.__dict__.keys()):
-        if key.startswith("__"):
-            continue
-        if key.endswith("__"):
-            continue
-        delattr(models, key)
-
-
 @pytest.fixture
 def mocked_facades_models(monkeypatch):
     """Monkeypatch open_alchemy.facades.models."""
@@ -167,3 +133,23 @@ def mocked_facades_sqlalchemy(monkeypatch):
     mock_sqlalchemy = mock.MagicMock()
     monkeypatch.setattr(open_alchemy.facades, "sqlalchemy", mock_sqlalchemy)
     return mock_sqlalchemy
+
+
+@pytest.fixture(autouse=True)
+def cleanup_models():
+    """Remove any new attributes on open_alchemy.models."""
+    for key in set(models.__dict__.keys()):
+        if key.startswith("__"):
+            continue
+        if key.endswith("__"):
+            continue
+        delattr(models, key)
+
+    yield
+
+    for key in set(models.__dict__.keys()):
+        if key.startswith("__"):
+            continue
+        if key.endswith("__"):
+            continue
+        delattr(models, key)